+++ CHANGELOG.rst
@@ -17,6 +17,10 @@ Fixed
 ^^^^^
 * /etc has the lowest priority and the current working directory has the highest
 
+Changed
+^^^^^^^
+* Rename FileReader to SettingsParser
+
 `2.1.0`_ - 2022-04-28
 ---------------------